ログイン・サインアップページの共通化
